Responsive Design
All content is responsive and displays on a range of display sizes.

Content should make use of available screen real estate and should display correctly at all screen sizes.

An image's associated title and text renders next to the image in all viewport sizes.

Images in the site are sized appropriate to the viewport and do not crowd or overlap other elements in the browser, regardless of viewport size.

On the main page, restaurants and images are displayed in all viewports. The detail page includes a map, hours and reviews in all viewports.